5 задание. Перенес getArr в data.js
