+++ js/utilits.js
@@ -1,4 +1,3 @@
-import {data} from './data.js';
 // Функция, возвращающая случайное целое число из переданного диапазона включительно.
 const getRandomInt = function (min, max) {
   return Math.abs(Math.floor(Math.random() * (max - min)) + min);
@@ -19,37 +18,8 @@ const rndArrey = function (arr) {
   }
   return rndArr;
 };
-
-const getArr = function (length) {
-  const arr = [];
-  const rndAdress = [];
-  for (let i = 0; i < length; i++) {
-    rndAdress[0] = getRandomFloat(35.65, 35.7, 5);
-    rndAdress[1] = getRandomFloat(139.7, 139.8, 5);
-    arr[i] = {
-      author: {
-        avatar: `img/avatars/user${((i + 1 < 10) ? (`0${i + 1}.png`) : (`${i + 1}.png`))}`
-      },
-      offer: {
-        title: 'Заголовок',
-        address: `${rndAdress[0]}, ${rndAdress[1]}`,
-        price: getRandomInt(15000, 25000),
-        type: data.typesOfPalaces[getRandomInt(0, 4)],
-        rooms: getRandomInt(1, 15),
-        guests: getRandomInt(1, 125),
-        checkin: data.times[getRandomInt(0, 2)],
-        checkout: data.times[getRandomInt(0, 2)],
-        features: rndArrey(data.features),
-        description: 'Описание помещения',
-        photos: rndArrey(data.photos),
-      },
-      location: {
-        lat: rndAdress[0],
-        lng: rndAdress[1],
-      },
-    };
-  }
-  return arr;
+export {
+  getRandomInt,
+  getRandomFloat,
+  rndArrey
 };
-const newArr = getArr(data.ARR_LENGTH);
-export {newArr};
